requirements.txt
numpy
sentence-transformers
pinecone-client
openai
# faiss dependency should be installed as per extras in setup.py